Move GetLongPathName to test helpers
