+++ internal/fileutils/fileutils_win.go
@@ -7,7 +7,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"syscall"
 
 	"github.com/ActiveState/cli/internal/logging"
 	"github.com/gobuffalo/packr"
@@ -86,23 +85,4 @@ func isWritableTempFile(path string) bool {
 	}
 
 	return true
-}
-
-// GetLongPath name returns the Windows long path (ie. ~1 notation is expanded)
-func GetLongPathName(path string) (string, error) {
-	p := syscall.StringToUTF16(path)
-	b := p // GetLongPathName says we can reuse buffer
-	n, err := syscall.GetLongPathName(&p[0], &b[0], uint32(len(b)))
-	if err != nil {
-		return "", err
-	}
-	if n > uint32(len(b)) {
-		b = make([]uint16, n)
-		n, err = syscall.GetLongPathName(&p[0], &b[0], uint32(len(b)))
-		if err != nil {
-			return "", err
-		}
-	}
-	b = b[:n]
-	return syscall.UTF16ToString(b), nil
 }
