+++ internal/testhelpers/osutil/path_helpers_lin_mac.go
@@ -1,0 +1,9 @@
+// +build !windows
+
+package osutil
+
+// GetLongPath name returns the Windows long path (ie. ~1 notation is expanded)
+// This function does not alter the path in any way on Linux or MacOS
+func GetLongPathName(path string) (string error) {
+	return path, nil
+}
